eos-fix-ostree-repo: Workaround broken tmp cleanup in old ostree
In older OSTree, cleaning up after a transaction (e.g., a pull) would
delete the tmp/cache directory if it was older than 1 day. That's a
problem because it has an open fd for that directory. Update the
directory's mtime to current. This is racy because other repo users may
have deleted the directory after we opened the repo and before they were
killed, so just fail if the directory doesn't exist.

https://phabricator.endlessm.com/T19420

squash! eos-fix-ostree-repo: Workaround broken cache directory in old ostree

https://phabricator.endlessm.com/T19420