eos-fix-ostree-repo: Make sure the repo is owned by root
Now that you can pass arbitrary repos to operate on, we want to make
sure that root owned files aren't created in non-root owned repos. It
would be nice to let this run unprivileged when possible, but that would
require kill_repo_procs() to handle that correctly. For now, just punt
if someone tries to run this on a repo that root doesn't own.

https://phabricator.endlessm.com/T18946