write-live-image: add ESP and bootable flag with --mbr
In principle, EFI systems are required to boot from devices with an MBR
rather than GPT. 0xEF is the MBR partition type ID for an ESP. Our ISOs
already have this structure, and do not have a GPT, and we are not
plagued with reports that they do not boot on EFI systems.

We know that some BIOS systems cannot boot from disks where no MBR
partition has the "bootable" flag set. So, set that too.