Add eos-dev-unlock script
This is like eos-dev-fix but it uses similar functionality inside
ostree to make /usr read-write.

It also acts differently, as any changes that you make are only
kept in RAM, so they are lost on reboot. This is quite convenient
for many development tasks. Thanks to Dan Nicholson for pointing
this out.

I'm leaving eos-dev-fix in place for now, as it may have other
interesting development use cases. I think I also saw once or twice
where apt was struggling to work with /usr as an overlayfs (which
does not quite match the behaviour of a real fs).