eos-diagnostics: reveal created file, be friendlier
Before:

    Gjs-Message: JS LOG: Saved to /sysroot/home/wjt/eos-diagnostic-171103_144228_UTC
    +0000.txt

(This message is too long for the default width of the terminal, even
with my short username; I have indicated how the line is wrapped in
practice.)

After:

    Endless OS Diagnostics file saved as:
    /sysroot/home/wjt/eos-diagnostic-171103_144325_UTC+0000.txt
    Please send that file (not this message!) to the Endless Community or Support.

and Nautilus pops up, with the file already highlighted.

I tried to use Gtk.RecentManager.get_default().add_item(filename) to add
it to the Recent tab in Nautilus and the file picker, but no dice.

https://phabricator.endlessm.com/T19833