eos-write-live-image: use MBR by default
We know there are BIOS machines which refuse to boot from the GPT
layout. We also know that recent versions of Windows mount all
partitions on the USB stick, including the (illegible) BIOS boot
partition, and helpfully offer to format them if they don't contain a
recognised filesystem. By switching to the MBR layout with only two
partitions -- both of which are comprehensible to Windows -- we avoid
both problems.

https://phabricator.endlessm.com/T19729