New eos-enable-coredumps command, reused it from eos-convert-system scripts
The new command fixes a problem in the current way coredumps were meant to
be enabled, which did not work because (1) eos-metrics-instrumentation would
override the kernel.core_pattern setting and (2) coredumps would be still
disabled due to the default 0-byte size limit set in limits.conf.

Besides, having a separate command makes it more convenient to enable
coredumps in non-converted systems.

https://phabricator.endlessm.com/T20331