eos-convert-system: Force update-grub
To protect users to run update-grub on ostree systems we have added a
check preventing update-grub from running when the 'ostree' string is
present in the cmdline. eos-convert-system is a special case since we
need to run it to regenerate the grub configuration when still running
an ostree system. Force update-grub passing the new (hidden) parameter
(-f | --force).

Signed-off-by: Carlo Caione <carlo@endlessm.com>