Make xserver-xorg-video-armsoc platform dependent
At least for the rk3288 image we do not have to install the armsoc
driver since this is breaking Xorg. For now we just want to fallback to
the modesetting driver.

Instead of unconditionally install xserver-xorg-video-armsoc on all the
armhf platforms, move this dependency in the platform dependent files so
that we can have a finer control where the package is being installed.

https://phabricator.endlessm.com/T21539