remove sorting of metapackage dependencies
We found a case where aspell-en was being installed un-necessarily
due to this dependency chain:
 evolution -> webkit -> enchant -> aspell-en
Even though enchant actually depends on:
aspell-en | hunspell-dictionaries | ...
And we otherwise install some providers of hunspell-dictionaries.

It seems that because evolution is evaluated by apt before hunspell-*
(versus being evaluated as part of eos-langpack-* previously) then
aspell-en is added un-necessarily.

Disable sorting of the metapackage dependencies and move hunspell-*
up in the list so that aspell-en isn't installed.