eos-stage-flatpak: Remove outdated appstream and ostree-metadata
When a flatpak remote is switched between the stages, the appstream data
and ostree-metadata become invalid, and updating them might fail if the
version in the new repo is older than the version you have (since
flatpak considers it a downgrade). So this commit forcibly deletes the
appstream data and ostree-metadata and fetches them anew after changing
the remote URLs.

https://phabricator.endlessm.com/T22314