eos-uninstall-dualboot: use python3 not python (2)
We no longer include python 2 in the OS, but this dependency was not
declared in the package metadata.

This script is now obsolete for all but extremely old installations, but
we keep it around since it is small and referenced in our documentation
(eg https://support.endlessm.com/hc/en-us/articles/214475283).

https://phabricator.endlessm.com/T22670