eos-diagnostics: Add a section for /ostree/repo/config
Since we seem to be getting more bug reports which concern incorrect
OSTree repository configurations, it’s probably useful to include the
contents of /ostree/repo/config in the diagnostics report.

Note that this means our repository passwords will be included in the
report, but since users can read them out of /ostree/repo/config anyway,
that’s probably not a problem.

Sample output:
  ===================================
  = OSTree repository configuration =
  ===================================
  [core]
  repo_version=1
  mode=bare
  min-free-space-percent=0
  add-remotes-config-dir=false
  …

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22886