eos-write-installer: fix calculating exFAT partition start
Since util-linux 2.32, sfdisk detects when the secondary GPT is not at
the end of the disk, and adjusts the header in the output of --dump
accordingly so that if you run `sfdisk --dump $DEVICE | sfdisk $DEVICE`
the GPT is extended to cover the entire device.

https://github.com/karelzak/util-linux/commit/9a8a3b88dcaec36d694aaea29d6fce5856d697c4

While this is helpful in general, it means we cannot use the last-lba
header to work out where the exFAT partition should start. Instead, we
can calculate it from the start and size of the final partition.

I got this wonderful regular expression from
https://github.com/endlessm/eos-boot-helper/blob/a4dee04/dracut/repartition/endless-repartition.sh#L163

https://phabricator.endlessm.com/T22943
(cherry picked from commit 39ea050a0a356ba29454a5ddd0384c2d6928b125)
Signed-off-by: Will Thompson <wjt@endlessm.com>