eos-stage-flatpak: remove appstream2 refs when changing remote
Newer versions of Flatpak generate two appstream refs on the remote, one which
is uncompressed so that deltas can be generated more successfully. To ensure
that Flatpak doesn't get stuck with the wrong metadata (or collection ID) in
the local repo, make sure both the old and the new refs are deleted.