eos-fix-mbr-swap-removal: Don't bail early if `sfdisk -d` fails
When we fail to dump the partition table from a disk we should simply
skip it and continuing iterating on the disks list.

I've also added a "welcome message" telling the user were to find logs
for the current session in case something else returns an error and
`bash -e` bails us out early.

https://phabricator.endlessm.com/T24493