Check that deps exist at build time
This would have prevented me from merging
https://github.com/endlessm/eos-meta/pull/554 when fuse-overlayfs was
not in the distro.

I knew keeping the Camel Book around would be useful!

When OBS builds packages, no apt cache is available, so this check
cannot run. It's mainly intended to be run by Jenkins for PRs, which
builds the dist tarball using "meson" + "ninja" rather than the Debian
tooling, so we can make this check optional, enable the option by
default, and disable it in debian/rules.

https://phabricator.endlessm.com/T26007