eos-dev-recommends: drop eos-dev-config
This package has not been installable for a year, makes surprising
changes to the current user's dotfiles, and we are talking about
potentially dropping it from the distro. I don't think the eos-dev
metapackage is used, either, but let's remove it from recommends here to
unblock checking all packages referenced by eos-meta are installable.