eos-dev-suggests: add ubuntu-keyring
Dan readded this to OBS after a previous deletion:

Nothing in Endless actually needs this, but some ancillary tools do. For
instance, if you want to run obs-add-package or jenkins-infra locally,
you'll need to be able to trust the Ubuntu repos. It has no additional
dependencies and won't be pulled into the OS.