eos-convert-system: Find kernel and initramfs in new location
We've updated our kernel location to match OSTree's suggested layout,
that means the kernel isn't where this script expected it to be anymore.

This isn't certain to work as intended if there are multiple kernel module
directories, so if a developer does a bunch of non-deterministic things
like eos-dev-unlock --hotfix and installing a new kernel before running
eos-convert-system, they may not end up booting the expected kernel.

https://phabricator.endlessm.com/T27124