Remove xapian-bridge
Removes xapian-bridge from the list of dependencies of eos-core, so it
should no longer be included in the ostree. This will render all SDK1
apps nonfunctional, but we believe that there are no longer any in use,
and if there are we will port them.

https://phabricator.endlessm.com/T27277
(cherry picked from commit 8e0e82cd6c22cf244fc0eb22e73ad0134d843224)