Rework S905X metapackage
We're reviving our Amlogic S905X R&D efforts using mainline
u-boot and mainline kernel.

We're still sticking to armhf as the base arch (as that's the only
arch we can support for the Mali GLESv2 driver), but need to include
arm64 packages for the kernel and bootloader.

eos-core-s905x is still a good name for the metapackage, because the
packages shipped are not limited to a single specific product based
around S905X.

However, the ostree builder will need to use a product-specific
devicetree, so will use per-device platform names. Since we're starting
with the libretech_cc platform, add an alias as eos-core-libretech_cc which
is what the ostree user will use.