eos-write-live-image: Allow provisioning of persistent storage space
We now have a command line option to fill the eoslive partition with
a file for persistent storage space.

Currently this is not a fast operation - truncating the file to size
appears to be slow, yet not quite as slow as zero filling the space
would be.  It is hoped that this will be faster in the future with
a non-FUSE exfat driver.  In that case we can reconsider whether
this mode of operation should be default.

Untested on NTFS.

https://phabricator.endlessm.com/T14891