eos-write-live-image: Invert --persistent-size
The goal of this parameter was to allow a certain amount of disk space
to be left free for changes to the Kolibri homedir.  However this means
that you have to know what the free space on the partition will be
before you run the command that populates the partition.

Instead, accept a --free-space parameter, and use the maximum of this
value (if provided) and 17 MB as the free space. This allows the caller
to directly request what they want: persistent.img filling all but X of
the free space.

https://phabricator.endlessm.com/T30677
(cherry picked from commit 68f6bd20075743b00aa5e7bcef7e2fd28ea5853b)