eos-write-live-image: Never create BIOS Boot partition
Because Windows offers to format the BIOS Boot partition, we don't want
to create it, hence --no-bios.  However, previously we didn't write a
BIOS bootloader when --no-bios was passed, which means the resulting USB
stick can only be booted on UEFI systems.

Even if we don't have an entry in the partition table, we can leave a
gap where the partition ought to be, and write the bootloader there
anyway.  This allows the resulting USB stick to be booted on a legacy
BIOS system too.

Continue to accept --no-bios, because we have documentation on creating
Hack Keys that mentions this parameter, but ignore it.

https://phabricator.endlessm.com/T30692