eos-write-live-image: Match ISOs by filename
On an Ubuntu 20.04 Focal VM, `file --mime-type` reports
'application/octet-stream' for identical bytes that are reported as
'application/x-iso9660-image' on my Endless OS system.

Life is too short for this. Simplify the check by moving it below the
'check_exists' invocation for the $OS_IMAGE (there are no accesses to
$ISO between these two points) and match the filename with a glob.

https://phabricator.endlessm.com/T30765