eos-write-live-image: Don't use udisks
I wrote this using udisks because I wanted to avoid the script having to
manually manage mountpoints. It also meant it could be run by an
unprivileged user if they can write to the target directory.

However:

- The logic to copy the squashfs out of an ISO just calls mount
  directly, meaning the script must now be run as root anyway.
- Talking to udisks is less appropriate when running this on a server;
  in particular I just tried to run it in a minimal Ubuntu 20.04 VM and
  failed.

This patch rewrites the logic for the other mount pounts to just call
'mkdir' and 'mount'. An idempotent cleanup function is registered to be
called on unexpected exit.

https://phabricator.endlessm.com/T30765