eos-write-live-image: Write partitions in physical order
We have historically had the last physical partition on the disk, the
image partition, be the first logical partition on the disk: in the
past, Windows would only mount the first logical partition on removable
media.

Some time during the life of Windows 10, this behaviour changed, and it
now tries to mount all partitions on the disk. So we no longer need to
do this trick.

As a side-effect of this change, `eos-live-storage-setup` in
`eos-boot-helper` will no longer attempt to create an additional
partition for persistent data, since it only does so when the disk image
is on partition 1. This is desirable because, in the previous
arrangement, it would create a 1MiB ext4 partition in the gap we have
carefully left to store the bootloader data in.

https://phabricator.endlessm.com/T30692