Pass '--wipe always' to sfdisk
I think this the warning pasted below, when writing disks that
previously had an ISO written to them -- needs testing.

The device contains 'iso9660' signature and it may remain on the device. It is recommended to wipe the device with wipefs(8) or sfdisk --wipe, in order to avoid possible collisions.