eos-initramfs-efi: Remove metapackage
This was used when the payg blob was built as part of the kernel build,
and we needed to ensure the right packages were loaded for dracut's
invocation at that time.

However, we now build the initramfs in eos-ostree-builder, after the
rest of the system is assembled and all required packages are present.

Nothing depends on eos-initramfs-efi anymore, so remove it.

https://phabricator.endlessm.com/T30878