core-arm64-depends: Add xserver-xorg-legacy dependency to arm64 OSTree
The xserver-xorg-legacy package provides the /usr/lib/xorg/Xorg.wrap
suid binary wrapper which wraps the regular Xorg binary and, when KMS is
not available, opens the fb device as root, drops privileges and execs
X, passing the open fds. It only allows users logged in to a console to
run X as root by default, which should prevent any remote exploits.

[commit message by jprvita]

https://phabricator.endlessm.com/T30918