eos-core: remove imagescan Epson scanner driver
Unfortunately it has been difficult to keep this package building and
working over the years, and it is now in broken state once again (after the
work in T31261).

With our refined focus I'm not convinced it's worth us continuing to put up
with the pain associated with this awkward device support, so I am removing
it from EOS4.

https://phabricator.endlessm.com/T11319