eos-download-image: Don't require --size with --url
This was added to work around a CloudFront limitation: you cannot
download more than 20 GiB in one go, nor can you even issue a HEAD
request to find out the file size.

Fastly (and for that matter Bunny) does not have this limitation, so we
can remove this artificial restriction.

(We could go further and remove all the logic that downloads files in
20 GiB chunks, but that code is harmless.)

https://phabricator.endlessm.com/T23134