Explicitly depend on wget
This used to be pulled in as a (transitive) dependency of something
else. (I'm not actually sure what.)

I think it is reasonable to include at least one of 'curl' and 'wget'.
Since we have historically shipped 'wget', and I found recent posts on
the forum assuming it was present, let's keep it.

https://phabricator.endlessm.com/T31889