eos-stage-ostree: Remove branch handling
Previously the development OS repo would not have reliable stable refs
because there might be multiple active branches publishing the stable
ref at the same time. For example, when `eos3.X` is branched but not yet
released, there are likely still builds for `eos3.X-1` happening and
both builds would be pushing to the `eos3a` stable branch ref. The same
goes for the new `latestN` branch ref.

Now the ostree builder configurations have been cleaned up so that only
one build at a time should be publishing a given stable or latest
branch. With that we can reliably use the same branch in any repo stage
and don't need to use heuristics to try to change it. `eos-stage-ostree`
is changed to simply update the repo URL and collection ID and leave the
branch alone. Anyone switching between OS refs is doing something fairly
advanced and can handle it outside of this script.

https://phabricator.endlessm.com/T32639