eos-core: Readd gnome-menus
In previous branches we patched the gnome-shell package to depend on
this. We (correctly IMO) dropped that patch this cycle.

This package provides many files in /usr/share/desktop-directories/
which correspond to standard app categories. GNOME Shell uses these to
suggest a default name for a folder when one app is dragged onto
another, by finding categories the two apps have in common and looking
up the localized name from a .directory file.

https://phabricator.endlessm.com/T28366
https://phabricator.endlessm.com/T32975