Remove flatpak-builder
We take flatpak-builder directly from Debian, but maintain our own
Flatpak. Due to a security issue, flatpak-builder now has a versioned
dependency on a newer version of Flatpak than we have in the distro.

Remove flatpak-builder. We can readd it once we have updated Flatpak. In
the meantime the org.flatpak.Builder package from Flathub can be used
instead.

https://phabricator.endlessm.com/T32961