core: Replace transitional ttf-ancient-fonts
This package is empty and its description states:

> This dummy package is provided for a smooth transition from the
> previous ttf-ancient-fonts package to the fonts-ancient-scripts package.

Actually, on closer inspection, it also depends on fonts-symbola.

(Why do we ship 42 MB of fonts for ancient scripts? It was added for the
History app, which has an article about Cuneiform, which requires the
Akkadian font. Why are these in the OS rather than being bundled in the
app that needs them? Good question! https://phabricator.endlessm.com/T27023)