mimeapps: Dissociate ostree-repos from GNOME Software
Previously, inserting a USB drive with an OSTree repo at its root would
cause Nautilus to prompt it to be opened with "Software Install". This
used to be correct, but when updating to GNOME Software 41 we dropped
our downstream support for USB updates.

This spiritually reverts 044c635e ("defaults: Associate GNOME Software
with ostree repos") and can be in turn reverted as and when Software
supports USB OS & app updates again.

https://phabricator.endlessm.com/T33546