os: Depend on pulseaudio-utils
This was previously installed as a dependency of pulseaudio itself, but
pulseaudio was removed in b864e88191a46c0bd7fe6f398aa20d2569fcb8ac
("os-depends: Switch to PipeWire for audio").

But apparently pipewire-pulse attempts to invoke pactl, logging an error
to the journal when it cannot be found. The pipewire-pulse package only
Suggests pulseaudio-utils. While the error does not appear to be fatal,
it adds noise to the journal, easily avoided by just installing this
small package.

https://phabricator.endlessm.com/T34099