Add udev rules for Android devices
The 'adb' tool can be installed in Toolbox.  If you use it to try to
debug an Android device over USB:

- First it will complain that your user is not in the 'plugdev' group.
  You need to run 'sudo adduser $USER plugdev' within the toolbox
  container to defeat this check; even though this makes no difference
  to the next step.
- Then it will fail to connect due to lack of permission to write to the
  USB device.

This package contains a load of udev rules that match Android devices
with USB debugging enabled connected via USB, and make them writable by
users in the 'plugdev' group. With this installed, you can add your real
user account (on the host system) to plugdev, and then successfully
connect using 'adb' within toolbox.

For the vast majority of Endless OS users who are not trying to debug
their Android phones, this is harmless, and only adds 32 KiB.

https://phabricator.endlessm.com/T33130