Remove gstreamer1.0-vaapi from OS
We do not ship VAAPI drivers in the OS.

Previously, we added symbolic links to pick up the drivers from a
Flatpak, if installed, but we are removing these because:

1. The drivers stopped working on the host for unknown reasons
2. Very little in the base OS uses GStreamer anyway

So there is no need to ship these libva-based GStreamer elements. They
never previously worked anyway, because we previously shipped a
systemwide GStreamer element cache which was precomputed on our build
server which did not have VAAPI drivers installed, and was not
recomputed if the VAAPI drivers were installed client-side.

https://phabricator.endlessm.com/T32994
https://phabricator.endlessm.com/T34998