eos-write-installer: Run as a regular user
This program still needs superuser access to format a removable device,
but it is able to run as a regular user to copy OS image files to the
eosimages partition. By only running specific privileged commands as
root, we can make it easier to use. In particular, with this change the
program can continue as usual if the eosimages partition has already
been mounted by the desktop user.

https://phabricator.endlessm.com/T35167