eos-diagnostics: Collect extra PAYG logs
eos-paygd writes extra logs to /var/log/eos-payg. These are useful on
systems where the journal is not persistent; and when we want deeper
history of PAYG state changes after the journal has been rotated.

I tested the following error cases:

- /var/log/eos-payg does not exist or is not readable; in this case the
  section is omitted. This should not happen in practice – the
  tmpfiles.d snippet sets mode 755.
- A file within is not readable; in this case the exception is included
  in the diagnostic file.
- A file within contains invalid UTF-8; in this case the replacement
  character is used.

If the directory is empty, the function returns empty string, which
means the section is omitted.

https://phabricator.endlessm.com/T33896