Revert "Revert "Remove gstreamer1.0-libav from OS""
This reverts commit 9439f13f3b723f128fac99fde39a41ee25287c24. For real
this time. The only other user of ffmpeg was freerdp, and it was
actually useless since our ffmpeg doesn't have H.264 codecs. So, freerdp
has been changed to not build against ffmpeg and now gstreamer1.0-libav
is the last user of it. After this, EOS will be free of ffmpeg.

https://phabricator.endlessm.com/T35225