eos-write-installer: Fix finding eosimages partition after repartitioning
Prior to commit ce670ad61b169bcf403b641e0b55a3e3139efb36
(“eos-write-installer: Run as a regular user”), after rewriting the
partition table we would parse the output of `sfdisk --dump` to find the
device node for the eosimages partition, in order to format it as exFAT.
Since `sfdisk` reads the partition table directly from the disk, this
doesn't depend on the kernel & udev having finished probing the new
partition layout and recreating the various device nodes.

In that commit, this was replaced by parsing the output of `lsblk`. On
my system, the result was always that lsblk would output no information
about the partitions, only the following:

    /dev/sdb

This appears to have been because we would run `partprobe` but not
follow it with `udevadm settle` to wait for the probing & recreation of
device nodes to finish.

As a result the script would set $EOSIMAGES_PART to empty string; skip
the mkfs.exfat due to the `if [ -n "$EOSIMAGES_PART" ]` check; and then
fail because empty string is not a block device.

Fix this by invoking `udevadm settle` after invoking `partprobe`, in
addition to invoking `udevadm settle` before invoking `partprobe`.

https://phabricator.endlessm.com/T35362